submit_for_testing: add support for multiple variables ini file
It might be convenient to store variables in multiple files instead of
one big file. So let --variables takes more than 1 file, and process
them *in order* so that if the same variable is defined in multiple
files the last value wins.

For example it allows to keep a generic file used by every project,
and then each project (Schneider, QCOM) can maintain their own file
with their custom variables , or overrides.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>