submit_for_testing: Rename --build-number to --build-id
It's a more general case of having git-describe instead of
build numbers in QA Reports as the build ID. It is up to the
user to name their builds with whatever scheme suites them,
so this allows them to user a build number, a git-describe,
a patronymic, or random name.

For now, both --build-number and --build-id will coexist, but
the latter will go extinct once we cut its supply of bits in
about 2-months' time.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>