android: add testcase templates for x15 and hikey960
with the docker method used for lava jobs.
needs to specify the --testplan-path to testcases/android,
and --testplan-device-path to devices/android/,
for the submit_for_testing.py script to be run.

Here created links for template-xxxx.yaml files, because
the current build configs use filename like this.
Will remove these template-xxx.yaml files when the build configs
side changed

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>