timeout calculation: add the (pre|post)_boot_timeout
Like this job here:
    https://lkft.validation.linaro.org/scheduler/job/5005860/definition#defline75

There are pre_os_command, pre_power_command, pre_os_command commands specified,
and there is a default timeout value(30s) for them but ignored,.

The commands are defined via the pre_boot_command and post_boot_command blocks
in the include/fastboot.jinja2 file, and it's OK to ignore them for this case
as the value is very little, but if time consuming jobs are defined via the
pre_boot_command and post_boot_command blocks, the job timeout value difference
will be obvious.

Hence this change to add the (pre|post)_boot_timeout variables
so that it's possible to specify the timeout values for the time consuming actions
when necessary.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>