add a fastboot_deploy_timeout variable
The job_timeout is calculated wrong when running fastboot jobs.
Only caclulate target_deploy_timeout once. Instead of calculate it twice
just split up the fastboot deploy with its own timeout variable
'fastboot_deploy_timeout'. Leave the first target_deploy_timeout as is
even if it does 'to: download' and not 'to: target'

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>