Create hip07-d05 board def and benchmark tests
Creating the hip07-d05 board definition for the device which is currently in the Cambridge Lab. This D05 device type specifically bypasses NFS and boots to local disk in the boot commands, as the purpose of these benchmark tests are to get accurate on board performance numbers.

Creating imagenet-ssd-resnet32.yaml, imagenet-resnet50.yaml and ssd-mobilenet.yaml benchmarks which run the corresponding tensorflow MLPerf benchmark from the test-definitions repository.

Signed-off-by: Theodore Grey <theodore.grey@linaro.org>