testcases: master: template-ltp.yaml.jinja2: set default TIMEOUT_MULTIPLIER to 10
With reference to recent change in upstream LTP
       0f72d519a Introduce a concept of max runtime

LTP each test timeout value reduced from 300 sec to 30 sec.
Setting up default time is to 30 sec to not adequate on
slow machines like 32-bit arch and qemu's.

The timeout part is a limit for the test setup and cleanup
and also safety margin for the runtime accounting. It's
currently set to 30 seconds but may change later. If your
target machine is too slow it can be scaled up with the
'LTP_TIMEOUT_MUL' environment variable.

For the slow machines we need atleast 5 minutes of timeout.
   TIMEOUT_MUL = 10.

Ref:
LTP commit ID:
0f72d519a Introduce a concept of max runtime

Signed-off-by: Naresh Kamboju <naresh.kamboju@linaro.org>