testcase: master: template-test: output results to '/dev/kmsg'
When a string is sent by the DUT is split in two due to a dmesg message,
so lava does not parse it, and that's why it fails later with the
message 'Test error: Invalid TESTCASE signal'.

Example of an error in a job:
'Received signal: <ST[>   24.020614] systemd-journald[301]: Oldest entry
in /run/log/jARTRUN 0_ltp-ipc 5650939_2.1.4.1'
The above signal LAVA will see the dmesg message as two different
messages one before and another after.

Enable output to be redirect the test shell messages to '/dev/kmsg', by
adding 'lava-signal: kmsg', inside each test-block.

See lava documentation [1] for more information.

[1] https://docs.lavasoftware.org/lava/test-repositories.html?highlight=kmsg#syntax
Suggested-by: Rémi Duraffort <remi.duraffort@linaro.org>
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>