include: test_target: support to specify docker image
When the docker image is not specified for the
test definitions, there is a default docker image will be used.
But for the android test jobs, it needs to use a customized
docker image instead of the default one.
like the test action defined here:
    https://lkft.validation.linaro.org/scheduler/job/5891482/definition#defline69

Hence this change to add the support for that.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>