lkft-android: fastboot: separate download and fastboot for deploy
As for Android jobs, the DUT might be rebooted during the test,
and the tests need to reboot the DUT with the boot.img
used before, that requires do the deploy action in two steps:
download first and then run fastboot with the downloaded
images. And this change is for that purpose.

The lava job generated would like this:
    https://lkft.validation.linaro.org/scheduler/job/5966988/definition

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>