boot_target: move most boot_target block into include
so that it could be reused with different variables passed.

But since there are some blocks defined inside the
boot_target block, which could not be moved to
the include file, so just most of the lines are moved to
the include file, thus necessary lines like the "- boot:"
line should be added as well when reused in other places.

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>