include: fastboot: boot: override kernel-start-message
While booting HW like e850-96, LAVA failes to match the kerel boot msg
['Linux version [0-9]'], When that happens the auto-login-action fails
which means that the job fails.
By changing the string that LAVA should search for
'kernel-start-message' to something that actually shows up in the log
makes the auto-login-action pass and the kernel boots up fine.
This is only true if the kernel_start_message variable is set to a
string for LAVA to match.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>