testcases: master: template-master: make EXTRA_KERNEL_ARGS an external variable
Tried to change the extra_kernel_args in the LAVA job for a specific
job, this could be done by setting the 'extra_kernel_args' variable. In
order to set an externa input variable this should be done from upercase
variables.
Change the 'extra_kernel_args' variable to be 'EXTRA_KERNEL_ARGS'.

Signed-off-by: Anders Roxell <anders.roxell@linaro.org>