overlay: Implement overlay option
The overlay option allows to supply any number of overlays with the
destination path, where the overlay has to be applied. If the
destination path is not given the default will be "/".

This avoids hard coding overlay parameters as part of the jinja
templates. We are not removing the existing overlay parameters from
the jinja templates in order to not break the backwards compatibility
for users who are already relying on this.

Signed-off-by: Senthil Kumaran S <senthil.kumaran@linaro.org>